set debug.hwui.render_dirty_regions=false
The IMG architecture does not support the EGL_SWAP_BEHAVIOR_PRESERVED_BIT
in EGL configurations. This bit is used only to be able to change the
EGL_SWAP_BEHAVIOR, which matters when the default behavior is not the
one we want. However on IMG, the default swap behavior is the one we
desire.

Conflicts:
	device.mk

Change-Id: Ice475770671bb1ccc54c3786c69b7ead5b48760f